requirements.txt
ultralytics>=8.0.0
opencv-python>=4.7.0
numpy>=1.21.0
pandas>=1.3.0
streamlit>=1.22.0
pillow>=9.0.0
matplotlib>=3.5.0
pyyaml>=6.0.0
